% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atck-cdf.R
\name{attck_cdf_tactic}
\alias{attck_cdf_tactic}
\title{Product an ATT&CK Cumulative Distribution Function by Tactic}
\usage{
attck_cdf_tactic(xdf, input = NULL, output = NULL, matrix = NULL, ...)
}
\arguments{
\item{xdf}{a data frame with \code{tactic}, \code{technique} and \code{value} columns.
If no \code{value} column exists, then the function will assume you
have passed in individual events and will perform a "count"
summarization before generating the heatmap.}

\item{input, output, matrix}{if both are not \code{NULL} then they should be
what \code{\link[=fct_tactic]{fct_tactic()}} takes as parameters. Otherwise, the function
will assume that the \code{tactic} column is already an ordered factor.}

\item{...}{passed on to \code{\link[ggplot2:geom_label]{ggplot2::geom_label()}}}
}
\description{
Product an ATT&CK Cumulative Distribution Function by Tactic
}
