% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attck-map.R
\name{attck_map}
\alias{attck_map}
\title{Generate an ATT&CK heatmap}
\usage{
attck_map(xdf, input = NULL, output = NULL, matrix = NULL,
  tile_col = "white", tile_size = 0.5, dark_lab = "white",
  light_lab = "black", dark_value_threshold = NULL, ...)
}
\arguments{
\item{xdf}{a data frame with \code{tactic}, \code{technique} and \code{value} columns.
If no \code{value} column exists, then the function will assume you
have passed in individual events and will perform a "count"
summarization before generating the heatmap.}

\item{input, output, matrix}{if both are not \code{NULL} then they should be
what \code{\link[=fct_tactic]{fct_tactic()}} takes as parameters. Otherwise, the function
will assume that the \code{tactic} column is already an ordered factor.}

\item{tile_col, tile_size}{color/size for the tile borders;
defaults to "\code{white}" and \code{0.5}, respectively.}

\item{dark_lab, light_lab}{text colors for when they appear on top of a
dark or light tile}

\item{dark_value_threshold}{since you can supply your own fill scale
and may use a transformation (e.g. "\code{log10}") when doing so, you
can specify the cutoff value for when to use \code{dark_lab} vs \code{light_lab}.
If \code{NULL} then half of \code{max(value)} will be used.}

\item{...}{passed on to the internal call to \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}}
}
\value{
a ggplot2 plot object which you can add a fill scale to as well
as themeing.
}
\description{
Generate an ATT&CK heatmap
}
