% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attckr-package.R
\docType{package}
\name{attckr}
\alias{attckr}
\alias{attckr-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/attckr}
\item BugReports: \url{https://gitlab.com/hrbrmstr/attckr/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
