% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{enterprise_attack}
\alias{enterprise_attack}
\title{Enterprise Attack Taxonomy v2.0}
\description{
Enterprise Attack Taxonomy v2.0
}
\note{
Id: \verb{bundle--83dad14b-ae53-4473-9f95-5ae37c8eaa5d}

Last updated: 2020-06-17
}
\references{
\url{https://github.com/mitre/cti/raw/master/enterprise-attack/enterprise-attack.json}
}
