% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct-tactic.R
\name{fct_tactic}
\alias{fct_tactic}
\title{Make an ordered Tactics factor with optional better labelling}
\usage{
fct_tactic(tactics, input = c("id", "pretty", "nl"),
  output = c("pretty", "nl", "id"), matrix = c("enterprise", "mobile",
  "pre"))
}
\arguments{
\item{tactics}{a character vector}

\item{input}{what is in \code{tactics}?}

\item{output}{what do you want the factor label to be?}

\item{matrix}{which matrix?}
}
\description{
Make an ordered Tactics factor with optional better labelling
}
