% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{mobile_attack}
\alias{mobile_attack}
\title{Mobile Attack Taxonomy v2.0}
\description{
Mobile Attack Taxonomy v2.0
}
\note{
Id: \code{bundle--d22f39d8-4fa0-4557-a925-1d7bbaffaa46}

Last updated: 2019-10-24
}
\references{
\url{https://github.com/mitre/cti/raw/master/mobile-attack/mobile-attack.json}
}
