% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{pre_attack}
\alias{pre_attack}
\title{Pre-Attack Taxonomy v2.0}
\description{
Pre-Attack Taxonomy v2.0
}
\note{
Id: \code{bundle--0bf62d3b-a5a3-4dae-996d-e990ca6e2f4d}

Last updated: 2019-07-31
}
\references{
\url{https://github.com/mitre/cti/raw/master/pre-attack/pre-attack.json}
}
