% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{tactics_f}
\alias{tactics_f}
\title{Tactics factors (generally for sorting & pretty-printing)}
\description{
Tactics factors (generally for sorting & pretty-printing)
}
\note{
Last updated: 2020-06-17
}
