% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{tidy_attack}
\alias{tidy_attack}
\title{Combined ATT&CK Matricies Tactics, Techniques and Technique detail}
\description{
Combined ATT&CK Matricies Tactics, Techniques and Technique detail
}
\note{
Last updated: 2019-10-24
}
