% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-tactics.R
\name{validate_tactics}
\alias{validate_tactics}
\title{Validate Tactics strings against MITRE authoritative source}
\usage{
validate_tactics(tactics, matrix = c("enterprise", "mobile", "pre"),
  na_rm = TRUE)
}
\arguments{
\item{tactics}{a character vector of tactic strings to validate. This will be
converted to lower-case, left/right spaces will be trimmed and
internal spaces will be converted to a single \code{-}}

\item{matrix}{which matrix to use when validating?}

\item{na_rm}{remove NA's before comparing?}
}
\value{
\code{TRUE} if all tactics validate, otherwise \code{FALSE} with messages
identifying the invalid tactics.
}
\description{
Validate Tactics strings against MITRE authoritative source
}
\examples{
validate_tactics("persistence")
validate_tactics(c("persistence", "Persistence", "Persistance"))
}
