% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{pre_attack}
\alias{pre_attack}
\title{Pre-Attack Taxonomy v2.0}
\description{
Pre-Attack Taxonomy v2.0
}
\note{
Id: \verb{bundle--803f51fd-e986-493c-9ab1-0b33b42a4dec}

Last updated: 2020-06-17
}
\references{
\url{https://github.com/mitre/cti/raw/master/pre-attack/pre-attack.json}
}
